extract shared code to `Field`
