+++ packages/widgetbook/lib/src/fields/dropdown_field.dart
@@ -3,7 +3,6 @@ import 'package:widgetbook_core/widgetbook_core.dart';
 
 import '../state/state.dart';
 import 'field.dart';
-import 'field_codec.dart';
 import 'field_type.dart';
 
 class DropdownField<T> extends Field<T> {
@@ -23,34 +22,12 @@ class DropdownField<T> extends Field<T> {
   final String Function(T value) labelBuilder;
 
   @override
-  Widget build(BuildContext context) {
-    final state = WidgetbookState.of(context);
-    final groupMap = FieldCodec.decodeQueryGroup(state.queryParams[group]);
-    final value = codec.toValue(groupMap[name]);
-
-    // Notify change when field is built from new query params,
-    // to keep query params and locale state (e.g. addon's setting) in sync.
-    onChanged(context, value);
-
+  Widget buildField(BuildContext context, T? value) {
     return DropdownSetting<T>(
       options: values,
       initialSelection: value,
       optionValueBuilder: labelBuilder,
-      onSelected: (value) {
-        onChanged(context, value);
-
-        final newGroupMap = Map<String, String>.from(groupMap)
-          ..update(
-            name,
-            (_) => codec.toParam(value),
-            ifAbsent: () => codec.toParam(value),
-          );
-
-        state.updateQueryParam(
-          group,
-          FieldCodec.encodeQueryGroup(newGroupMap),
-        );
-      },
+      onSelected: (value) => updateField(context, value),
     );
   }
