+++ packages/widgetbook/lib/src/fields/field.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:widgetbook/src/fields/field_codec.dart';
+
+import '../state/state.dart';
+import 'field_codec.dart';
 import 'field_type.dart';
 
 abstract class Field<T> {
@@ -36,7 +38,38 @@ abstract class Field<T> {
 
   final void Function(BuildContext context, T? value) onChanged;
 
-  Widget build(BuildContext context);
+  Widget build(BuildContext context) {
+    final state = WidgetbookState.of(context);
+    final groupMap = FieldCodec.decodeQueryGroup(state.queryParams[group]);
+    final value = codec.toValue(groupMap[name]);
+
+    // Notify change when field is built from new query params,
+    // to keep query params and locale state (e.g. addon's setting) in sync.
+    onChanged(context, value);
+
+    return buildField(context, value);
+  }
+
+  Widget buildField(BuildContext context, T? value);
+
+  void updateField(BuildContext context, T value) {
+    final state = WidgetbookState.of(context);
+    final groupMap = FieldCodec.decodeQueryGroup(state.queryParams[group]);
+
+    final newGroupMap = Map<String, String>.from(groupMap)
+      ..update(
+        name,
+        (_) => codec.toParam(value),
+        ifAbsent: () => codec.toParam(value),
+      );
+
+    onChanged(context, value);
+
+    state.updateQueryParam(
+      group,
+      FieldCodec.encodeQueryGroup(newGroupMap),
+    );
+  }
 
   Map<String, dynamic> toFullJson() {
     return {
