+++ packages/widgetbook/lib/src/fields/number_field.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-import 'package:widgetbook/src/routing/router.dart';
 
 import 'field.dart';
 import 'field_codec.dart';
@@ -21,34 +19,13 @@ class NumberField extends Field<num> {
         );
 
   @override
-  Widget build(BuildContext context) {
-    final queryParams = GoRouterState.of(context).queryParams;
-    final groupMap = FieldCodec.decodeQueryGroup(queryParams[group]);
-    final value = codec.toValue(groupMap[name]);
-
-    // Notify change when field is built from new query params,
-    // to keep query params and locale state (e.g. addon's setting) in sync.
-    onChanged(context, value);
-
+  Widget buildField(BuildContext context, num? value) {
     return TextFormField(
       initialValue: codec.toParam(value ?? initialValue),
-      onChanged: (value) {
-        final number = codec.toValue(value) ?? initialValue;
-        onChanged(context, number);
-
-        final router = GoRouter.of(context);
-        final newGroupMap = Map<String, String>.from(groupMap)
-          ..update(
-            name,
-            (_) => codec.toParam(number),
-            ifAbsent: () => codec.toParam(number),
-          );
-
-        router.updateQueryParam(
-          group,
-          FieldCodec.encodeQueryGroup(newGroupMap),
-        );
-      },
+      onChanged: (value) => updateField(
+        context,
+        codec.toValue(value) ?? initialValue,
+      ),
     );
   }
